fix: sampling numbers went out as float noise
`temperature` and `top_p` were f32. `serde_json::Value` has no f32 variant, so
`json!(t)` widens to f64 and the f32 representation error becomes visible on
the wire: 0.7 is sent as 0.699999988079071, and 0.95 as 0.949999988079071.

Every provider has been receiving that. Most ignore the extra digits, which is
why it went unnoticed — until Z.AI, which validates to two decimal places and
rejects the request outright:

    "The temperature parameter is illegal.：限制小数点[2]位"

The fix is the type, not a rounding step. TOML parses floats as f64 already, so
the round trip was f64 -> f32 -> f64 and the narrowing was pure loss; f32 was
never buying anything here.

The test asserts the serialized body rather than the value, because the body is
what the provider actually parses.

+++ src/agent.rs
@@ -174,8 +174,8 @@ pub struct Agent {
     registry: Arc<ToolRegistry>,
     bus: EventBus,
     model: String,
-    temperature: Option<f32>,
-    top_p: Option<f32>,
+    temperature: Option<f64>,
+    top_p: Option<f64>,
     top_k: Option<u32>,
     max_tokens: Option<u32>,
     max_steps: usize,
@@ -200,7 +200,7 @@ impl Agent {
         registry: Arc<ToolRegistry>,
         bus: EventBus,
         model: String,
-        temperature: Option<f32>,
+        temperature: Option<f64>,
         max_tokens: Option<u32>,
         max_steps: usize,
         max_retries: usize,
@@ -237,7 +237,7 @@ impl Agent {
     /// Sampling the model asks for (`[models."provider/model"]`). Unset fields
     /// leave the server's own defaults alone, which is what every request did
     /// before this table existed.
-    pub fn with_sampling(mut self, temperature: Option<f32>, top_p: Option<f32>, top_k: Option<u32>) -> Self {
+    pub fn with_sampling(mut self, temperature: Option<f64>, top_p: Option<f64>, top_k: Option<u32>) -> Self {
         if temperature.is_some() {
             self.temperature = temperature;
         }
